requirements.txt
# AI 知识短视频素材引擎 - 依赖清单
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0.1
google-genai>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
python-dateutil>=2.8.2
